time

+++ GTServer/Server.cpp
@@ -935,13 +935,13 @@ namespace GT {
 		double timeInSeconds = 0;
 		clock_t endTime = clock();
 		for (std::map<SOCKET, GTClient>::iterator it = clients.begin(); it != clients.end(); ++it) {
-			timeInSeconds = (double(endTime) - double(it->second.clock) / CLOCKS_PER_SEC);
-
+			timeInSeconds = (double(endTime - it->second.clock) / CLOCKS_PER_SEC);
+			 
 			printf("%10d", it->second.header);
 			printf("%18s", it->second.address);
 			printf("%12s", it->second.name);
 			printf("%6d", it->second.header);
-			printf("%10.3f", timeInSeconds);
+			printf("%12.3f", timeInSeconds);
 			printf("%8d", int(it->second.socket));
 			printf("%8d", it->second.version_id);
 			printf("%6d\n", it->second.type);
@@ -952,7 +952,7 @@ namespace GT {
 				printf("%18s", it->second.address);
 				printf("%12s", it->second.name);
 				printf("%6d", it->second.header);
-				printf("%10.3f", timeInSeconds);
+				printf("%12.3f", timeInSeconds);
 				printf("%8d", int(it->second.socket));
 				printf("%8d", it->second.version_id);
 				printf("%6d\n", it->second.type);
